move some variables to variables.ts
